fix(Modal): generated docs use helper function as component
Move helper functions outside main component module
